fix(workflow): address Copilot review — optional timeoutMs, test relocated to tests/
- timeoutMs on PanSouFetchInit is optional with a shared
  DEFAULT_REQUEST_TIMEOUT_MS fallback in defaultFetchJson (the package is
  private so no external callers exist, but the defensive default costs
  nothing)
- timeout test moved into the existing tests/pansou-provider.test.ts
  (the package convention — provider tests live under tests/, and my
  earlier grep only checked src/); afterEach now closeAllConnections()
  and awaits server.close()

7 tests green in the merged file; 1097 package tests, build:workflow and
both tsc projects green.
